feat(#5): Added graphql code generator
* Now generates types from graphql schemas for web and server
* Added
getSchema script to combine multiple schemas
