+++ README.md
@@ -1,0 +1,10 @@
+## NOTE
+
+In test files absolute imports do not work. You need to use relative imports instead.
+
+```
+import app from "app"; // will not work
+
+import app from "./app"; // works fine
+import app from "../app"; 😊
+```
