+++ packages/common/tslint.json
@@ -1,5 +1,8 @@
 {
   "extends": ["../../tslint.json"],
+  "linterOptions": {
+    "exclude": ["build", "node_modules/**", "build/**"]
+  },
   "rules": {
     "no-console": false
   }
