+++ packages/server/src/resolvers/register/schema.graphql
@@ -1,7 +1,3 @@
-type Query {
-  hello(name: String): String!
-}
-
 type Mutation {
   register(email: String!, username: String!, password: String!): Boolean
 }
