+++ packages/server/tslint.json
@@ -1,5 +1,8 @@
 {
   "extends": ["../../tslint.json"],
+  "linterOptions": {
+    "exclude": ["node_modules/**", "build/**"]
+  },
   "rules": {
     "no-console": false
   }
